test/leetcode/ZeroOneMatrix_542Test.java
/*
 * Copyright (c) 2024 THGEEK.COM. All rights reserved.
 */

package leetcode;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

/**
 * Test for {@link ZeroOneMatrix_542}
 *
 * @author Tao Hong
 * @version 1.0
 * @since 2024/07/05 08:36
 */
class ZeroOneMatrix_542Test {
    private ZeroOneMatrix_542 test;

    @BeforeEach
    void setUp() {
        test = new ZeroOneMatrix_542();
    }

    @Test
    void test1() {
        int[][] mat = {{0, 0, 0}, {0, 1, 0}, {0, 0, 0}};
        int[][] expected = {{0, 0, 0}, {0, 1, 0}, {0, 0, 0}};
        int[][] actual = test.updateMatrix(mat);
        assertEquals(expected.length, actual.length);
        for (int i = 0; i < expected.length; i++) {
            assertArrayEquals(expected[i], actual[i]);
        }
    }

    @Test
    void test2() {
        int[][] mat = {{0, 0, 0}, {0, 1, 0}, {1, 1, 1}};
        int[][] expected = {{0, 0, 0}, {0, 1, 0}, {1, 2, 1}};
        int[][] actual = test.updateMatrix(mat);
        assertEquals(expected.length, actual.length);
        for (int i = 0; i < expected.length; i++) {
            assertArrayEquals(expected[i], actual[i]);
        }
    }
}